Add support for jiff, no_std, const and implement core::error::Error.
Jiff (https://github.com/BurntSushi/jiff) is a new date-time library
that is an alternative to `chrono`.

Everything in this crate already seem to support `no_std`, so I enabled
it.

Both `gregorian()` and `julian()` are now `const` since which allows
them to be used as static constants on compile time.

Returning strings makes the error messages harder to work with in
functions that automatically wraps the error (like `anyhow`). This
implements a proper type for it that implements `core::error::Error`.
This is however a fairly modern feature that requires at least Rust
`1.81.0`. I have therefore added `rust-version` to the package
specification.